requirements.txt
absl-py==1.0.0
asgiref==3.5.0
asn1crypto==1.5.1
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cssselect2==0.5.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.3
django-crispy-forms==1.14.0
django-filter==21.1
django-import-export==2.8.0
et-xmlfile==1.1.0
filelock==3.4.2
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
html5lib==1.1
idna==3.3
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupPy==1.14
nets==0.0.3.1
networkx==2.8
oauthlib==3.2.0
odfpy==1.4.1
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==21.3
Pillow==9.1.0
platformdirs==2.4.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyparsing==3.0.8
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Shapely==1.8.1.post1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.0
termcolor==1.1.0
tifffile==2022.4.8
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.1
wrapt==1.14.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
